+++ src/main/java/int222/integrated/Controllers/OnlineshopController.java
@@ -11,6 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import int222.integrated.Exception.ExceptionResponse;
+import int222.integrated.Exception.ProductException;
 //import int222.integrated.Exception.ExceptionResponse;
 //import int222.integrated.Exception.OnlineshopException;
 import int222.integrated.Models.Onlineshop;
@@ -30,17 +32,25 @@ public class OnlineshopController {
 	@GetMapping("/onlineshop/{onlineid}")
 	public Onlineshop showOnlineshops(@PathVariable int onlineid) {
 		Onlineshop onlineshop = this.onlineShopJpa.findById(onlineid).orElse(null);
+		if (onlineshop == null) {
+			throw new ProductException(ExceptionResponse.ERROR_CODE.ONLINE_ID_DOES_NOT_EXIST,
+					"onlineid id : " + onlineid + " does not exist ");
+		}
 		return onlineshop;
 	}
-	
-	@PostMapping(value = "/addOnlineshop")
+
+	@PostMapping(value = "/addonlineshop")
 	public Onlineshop create(@RequestBody Onlineshop newOnlineshop) {
 		return onlineShopJpa.save(newOnlineshop);
 	}
-	
+
 	@DeleteMapping("/onlineshop/{onlineid}")
 	public String delete(@PathVariable Integer onlineid) {
-		onlineShopJpa.findById(onlineid).orElse(null);
+		Onlineshop onlineshop = this.onlineShopJpa.findById(onlineid).orElse(null);
+		if (onlineshop == null) {
+			throw new ProductException(ExceptionResponse.ERROR_CODE.ONLINE_ID_DOES_NOT_EXIST,
+					"onlineid id : " + onlineid + " does not exist ");
+		}
 		onlineShopJpa.deleteById(onlineid);
 		return "Delete Online Shop  Success";
 	}
